# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  delivery.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
...
  delivery.UpdateGuestAttendingAndMessage:
    properties:
      is_attending:
        type: boolean
      message:
        type: string
      short_id:
        type: string
    type: object
...
  entity.Guest:
    properties:
      is_arrived:
        type: boolean
      is_invitation_sent:
        type: boolean
      is_vip:
        type: boolean
      message:
        type: string
      name:
        type: string
      phone_number:
        type: string
      qr_code_identifier:
        type: string
      short_id:
        type: string
      uuid:
        type: string
      will_attend_event:
        type: boolean
    type: object
...
host: gosm.muhammadilham.xyz
info:
  contact:
    email: hi@muhammadilham.xyz
    name: API Support
    url: https://pumbook.muhammadilham.xyz
  description: A guest management system that is written in Go.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GoSM
  version: "1.0"
paths:
...
  /api/v1/public/guests:
    get:
      consumes:
      - application/json
      description: Fetches guest information without requiring authentication.
      parameters:
      - description: Guest Short ID
        in: query
        name: short_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved guest
          schema:
            allOf:
            - $ref: '#/definitions/delivery.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.Guest'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery.Response'
        "404":
          description: Guest not found
          schema:
            $ref: '#/definitions/delivery.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery.Response'
      summary: Get guest by short ID
      tags:
      - guests
    post:
      consumes:
      - application/json
      description: Allows guests to update their attending status using a short ID.
      parameters:
      - description: Guest attending status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/delivery.UpdateGuestAttendingAndMessage'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated guest status
          schema:
            $ref: '#/definitions/delivery.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/delivery.Response'
      summary: Update guest attending status
      tags:
      - guests
...
swagger: "2.0"